+++ src/webR/webr-chan.ts
@@ -147,12 +147,30 @@ export interface FSMessage extends Message {
 }
 
 /** @internal */
+export interface FSAnalyzePathMessage extends Message {
+  type: 'analyzePath';
+  data: {
+    path: string,
+    dontResolveLastLink?: boolean,
+  };
+}
+
+/** @internal */
 export interface FSMountMessage extends Message {
   type: 'mount';
   data: {
     type: FSType;
     options: FSMountOptions;
     mountpoint: string;
+  };
+}
+
+/** @internal */
+export interface FSRenameMessage extends Message {
+  type: 'rename';
+  data: {
+    oldpath: string;
+    newpath: string;
   };
 }